requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
requests>=2.28.0,<3.0.0
beautifulsoup4>=4.10.0,<5.0.0
python-multipart>=0.0.5,<0.1.0
python-dotenv>=0.19.0,<0.20.0
qrcode>=7.0,<8.0
Pillow>=9.0.0,<10.0.0
